[Bugfix] fix drop hive table if exists (#31072)

+++ fe/fe-core/src/main/java/com/starrocks/connector/hive/HiveMetadata.java
@@ -147,6 +147,12 @@ public class HiveMetadata implements ConnectorMetadata {
                         " 'Force' must be set when dropping a hive table." +
                         " Please execute 'drop table %s.%s.%s force'", stmt.getCatalogName(), dbName, tableName));
             }
+
+            if (getTable(dbName, tableName) == null && stmt.isSetIfExists()) {
+                LOG.warn("Table {}.{} doesn't exist", dbName, tableName);
+                return;
+            }
+
             hmsOps.dropTable(dbName, tableName);
         }
     }
